+++ Common.Android/AndroidTicker.cs
