+++ Common.Android/StreamWrapper.cs
