+++ Common/Common.projitems
@@ -10,16 +10,35 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Color.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Device.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)EditorAutoSizeOption.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Font.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FontAttributes.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)IDispatcher.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Internals\IIsolatedStorageFile.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Internals\IPlatformServices.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Internals\Ticker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Log.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)LogListener.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)NamedPlatformColor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NamedSize.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NumericExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)OSAppTheme.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Point.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rect.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rectangle.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Size.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SizeRequest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SynchronizedList.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TextTransform.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TextType.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\ColorTypeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\PointTypeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\ProvideCompiledAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\RectangleTypeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\SizeTypeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\TypeConversionAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\TypeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Xaml\TypeConverterAttribute.cs" />
   </ItemGroup>
 </Project>
