+++ Common/EditorAutoSizeOption.cs
