+++ Common/IDispatcher.cs
