+++ Common/Internals/IIsolatedStorageFile.cs
