+++ Common/Internals/Ticker.cs
