+++ Common/NamedPlatformColor.cs
