+++ Common/Xaml/RectangleTypeConverter.cs
