+++ Common/Xaml/SizeTypeConverter.cs
