+++ System.Maui.Interfaces/IButton.cs
@@ -1,0 +1,9 @@
+﻿using Xamarin.Forms;
+
+namespace System.Maui
+{
+	public interface IButton : IText
+	{
+		void Clicked();
+	}
+}
