+++ System.Maui.Interfaces/IEditor.cs
@@ -1,0 +1,11 @@
+﻿using Xamarin.Forms;
+
+namespace System.Maui
+{
+	public interface IEditor : ITextInput
+	{
+		public EditorAutoSizeOption AutoSize { get; }
+
+		void Completed();
+	}
+}
