+++ System.Maui.Interfaces/ISearch.cs
@@ -1,0 +1,11 @@
+﻿using Xamarin.Forms;
+
+namespace System.Maui
+{
+	public interface ISearch : ITextInput
+	{
+		void Search();
+		void Cancel();
+		Color CancelColor { get; }
+	}
+}
