+++ Xamarin.Forms.Core/Device.cs
@@ -9,16 +9,8 @@ using Xamarin.Forms.Internals;
 
 namespace Xamarin.Forms
 {
-	public static class Device
+	public static partial class Device
 	{
-		public const string iOS = "iOS";
-		public const string Android = "Android";
-		public const string UWP = "UWP";
-		public const string macOS = "macOS";
-		public const string GTK = "GTK";
-		public const string Tizen = "Tizen";
-		public const string WPF = "WPF";
-
 		[EditorBrowsable(EditorBrowsableState.Never)]
 		public static DeviceInfo info;
 
@@ -72,11 +64,6 @@ namespace Xamarin.Forms
 		public static void SetFlowDirection(FlowDirection value) => FlowDirection = value;
 		public static FlowDirection FlowDirection { get; internal set; }
 
-		[EditorBrowsable(EditorBrowsableState.Never)]
-		public static bool IsInvokeRequired
-		{
-			get { return PlatformServices.IsInvokeRequired; }
-		}
 
 		[EditorBrowsable(EditorBrowsableState.Never)]
 		public static IPlatformServices PlatformServices
