+++ Xamarin.Forms.Platform.Android/Xamarin.Forms.Platform.Android.csproj
@@ -45,10 +45,6 @@
     <None Remove="Resources\values\attrs.xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Xamarin.Forms.Core\Crc64.cs" Link="Crc64.cs" />
-    <Compile Include="..\Xamarin.Forms.Core\StreamWrapper.cs" Link="StreamWrapper.cs" />
-  </ItemGroup>
-  <ItemGroup>
     <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Xml" />
